app/requirements.txt
sanic
python-dotenv
openai
httpx
sentence-transformers
hnswlib
numpy
